+++ shareHouse.xcworkspace/contents.xcworkspacedata
@@ -1,6 +1,10 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <Workspace
    version = "1.0">
+   <Group
+      location = "container:"
+      name = "portMOdel">
+   </Group>
    <FileRef
       location = "group:shareHouse.xcodeproj">
    </FileRef>
