+++ shareHouse/Home/AddView/AddViewController.swift
@@ -14,7 +14,5 @@ final class AddViewController: UIViewController {
     
     override func viewDidLoad() {
         super.viewDidLoad()
-        
-        
     }
  }
