--- shareHouse/Sigin/ViewController.swift
@@ -1,30 +1,0 @@
-//
-//  ViewController.swift
-//  shareHouse
-//
-//  Created by 奥田竜矢 on 2020/02/08.
-//  Copyright © 2020 okuda ryuya. All rights reserved.
-//
-
-import UIKit
-
-final class ViewController: UIViewController {
-    
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        
-        title = "Home"
-        setupNavigationController()
-    }
-    
-    private func setupNavigationController() {
-        let rigthButton = UIBarButtonItem(barButtonSystemItem: .add, target: self, action: #selector(self.presentAddView))
-        navigationItem.rightBarButtonItem = rigthButton
-    }
-    
-    @objc private func presentAddView() {
-        let addView = AddViewController.instantiate()
-        navigationController?.pushViewController(addView, animated: false)
-    }
-}
-
